- Registro de usuario (falta construcao da url e envio de email) - Validacao de registro de usuario pela url (validacao de novo registro e criacao de usuario ativo)

+++ src/br/usp/ime/ingpos/modelo/dao/CurriculoDAO.java
@@ -7,7 +7,7 @@ import br.usp.ime.ingpos.modelo.Curriculo;
 
 public class CurriculoDAO
     extends
-        AbstractDaoImpl<Integer,Curriculo>
+        AbstractDaoImpl<Long,Curriculo>
 {
 
     public CurriculoDAO(
